38.791 s -> detectOpenHandlesSlow.ts (5.011 s) detectOpenHandlesELD.ts (4.832 s)
